changes to respect ISO C99
git-svn-id: http://beaengine.googlecode.com/svn/trunk@51 e67b0f8a-f93e-11de-ba51-6b4196a2b81e

+++ beaengineSources/BeaEngine.h
@@ -290,6 +290,6 @@ enum SPECIAL_INFO
 #ifdef __cplusplus
 extern "C"
 #endif
-size_t __bea_callspec__ Disasm(LPDISASM pDisAsm);
+int __bea_callspec__ Disasm(LPDISASM pDisAsm);
 
 #endif
