+++ beaengineSources/Includes/instr_set/opcodes_Grp13.c
@@ -41,7 +41,7 @@ void __bea_callspec__ G13_(PDISASM pMyDisasm)
             if (!Security(0)) return;
             ImmediatSize = 8;
             MyNumber = *((UInt8*)(UIntPtr) (EIP_-1));
-            (void) CopyFormattedNumber((char*) &(*pMyDisasm).Argument2.ArgMnemonic,"%.2X", MyNumber);
+            (void) CopyFormattedNumber((char*) &(*pMyDisasm).Argument2.ArgMnemonic,"%.2X",(Int64) MyNumber);
             (*pMyDisasm).Instruction.Immediat = MyNumber;
             (*pMyDisasm).Argument2.ArgType = CONSTANT_TYPE+ABSOLUTE_;
             (*pMyDisasm).Argument2.ArgSize = 8;
@@ -62,7 +62,7 @@ void __bea_callspec__ G13_(PDISASM pMyDisasm)
             if (!Security(0)) return;
             ImmediatSize = 8;
             MyNumber = *((UInt8*)(UIntPtr) (EIP_-1));
-            (void) CopyFormattedNumber((char*) &(*pMyDisasm).Argument2.ArgMnemonic,"%.2X", MyNumber);
+            (void) CopyFormattedNumber((char*) &(*pMyDisasm).Argument2.ArgMnemonic,"%.2X",(Int64) MyNumber);
             (*pMyDisasm).Instruction.Immediat = MyNumber;
             (*pMyDisasm).Argument2.ArgType = CONSTANT_TYPE+ABSOLUTE_;
             (*pMyDisasm).Argument2.ArgSize = 8;
@@ -85,7 +85,7 @@ void __bea_callspec__ G13_(PDISASM pMyDisasm)
             if (!Security(0)) return;
             ImmediatSize = 8;
             MyNumber = *((UInt8*)(UIntPtr) (EIP_-1));
-            (void) CopyFormattedNumber((char*) &(*pMyDisasm).Argument2.ArgMnemonic,"%.2X", MyNumber);
+            (void) CopyFormattedNumber((char*) &(*pMyDisasm).Argument2.ArgMnemonic,"%.2X",(Int64) MyNumber);
             (*pMyDisasm).Instruction.Immediat = MyNumber;
             (*pMyDisasm).Argument2.ArgType = CONSTANT_TYPE+ABSOLUTE_;
             (*pMyDisasm).Argument2.ArgSize = 8;
@@ -106,7 +106,7 @@ void __bea_callspec__ G13_(PDISASM pMyDisasm)
             if (!Security(0)) return;
             ImmediatSize = 8;
             MyNumber = *((UInt8*)(UIntPtr) (EIP_-1));
-            (void) CopyFormattedNumber((char*) &(*pMyDisasm).Argument2.ArgMnemonic,"%.2X", MyNumber);
+            (void) CopyFormattedNumber((char*) &(*pMyDisasm).Argument2.ArgMnemonic,"%.2X",(Int64) MyNumber);
             (*pMyDisasm).Instruction.Immediat = MyNumber;
             (*pMyDisasm).Argument2.ArgType = CONSTANT_TYPE+ABSOLUTE_;
             (*pMyDisasm).Argument2.ArgSize = 8;
@@ -130,7 +130,7 @@ void __bea_callspec__ G13_(PDISASM pMyDisasm)
             if (!Security(0)) return;
             ImmediatSize = 8;
             MyNumber = *((UInt8*)(UIntPtr) (EIP_-1));
-            (void) CopyFormattedNumber((char*) &(*pMyDisasm).Argument2.ArgMnemonic,"%.2X", MyNumber);
+            (void) CopyFormattedNumber((char*) &(*pMyDisasm).Argument2.ArgMnemonic,"%.2X",(Int64) MyNumber);
             (*pMyDisasm).Instruction.Immediat = MyNumber;
             (*pMyDisasm).Argument2.ArgType = CONSTANT_TYPE+ABSOLUTE_;
             (*pMyDisasm).Argument2.ArgSize = 8;
@@ -151,7 +151,7 @@ void __bea_callspec__ G13_(PDISASM pMyDisasm)
             if (!Security(0)) return;
             ImmediatSize = 8;
             MyNumber = *((UInt8*)(UIntPtr) (EIP_-1));
-            (void) CopyFormattedNumber((char*) &(*pMyDisasm).Argument2.ArgMnemonic,"%.2X", MyNumber);
+            (void) CopyFormattedNumber((char*) &(*pMyDisasm).Argument2.ArgMnemonic,"%.2X",(Int64) MyNumber);
             (*pMyDisasm).Instruction.Immediat = MyNumber;
             (*pMyDisasm).Argument2.ArgType = CONSTANT_TYPE+ABSOLUTE_;
             (*pMyDisasm).Argument2.ArgSize = 8;
